requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-login
flask-bcrypt
python-dotenv
psycopg2-binary
pandas
numpy
scikit-learn
plotly
requests
pytest
